+++ src/lib.rs
@@ -9,7 +9,7 @@
 use std::fmt::Display;
 
 pub mod codes;
-pub mod maps;
+pub mod map;
 pub mod math;
 pub mod terms;
