+++ tests/lib.rs
@@ -1,5 +1,5 @@
 mod codes;
-mod maps;
+mod map;
 mod math;
 mod serialize;
 mod terms;
